refactor(risk-referral-service): Split riskAssessment and riskCondition routes out of controllers
Both features move route/middleware declarations and response Zod
schemas into their own *.routes.ts files; controllers keep only
request-handling logic via createXController(service) factories.
module.ts files now create the shared DocumentedRouter and register
routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after, lint clean, all 44 existing tests pass, and a live
smoke check against the running service.
