+++ apps/risk-referral-service/src/risk-assessments/riskAssessment.module.ts
@@ -1,15 +1,17 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { RiskAssessmentRepository } from './riskAssessment.repository';
 import { RiskAssessmentService } from './riskAssessment.service';
-import { createRiskAssessmentRouter } from './riskAssessment.controller';
+import { registerRiskAssessmentRoutes } from './riskAssessment.routes';
 
 /**
  * Composition root for the risk-assessments feature: wires repository →
- * service → router.
+ * service → routes.
  */
 export function createRiskAssessmentModule(prisma: PrismaService): DocumentedRouter {
   const repository = new RiskAssessmentRepository(prisma);
   const service = new RiskAssessmentService(repository);
-  return createRiskAssessmentRouter(service);
+  const doc = createDocumentedRouter();
+  registerRiskAssessmentRoutes(doc, service);
+  return doc;
 }
