+++ apps/risk-referral-service/src/risk-conditions/riskCondition.module.ts
@@ -1,15 +1,17 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { RiskConditionRepository } from './riskCondition.repository';
 import { RiskConditionService } from './riskCondition.service';
-import { createRiskConditionRouter } from './riskCondition.controller';
+import { registerRiskConditionRoutes } from './riskCondition.routes';
 
 /**
  * Composition root for the risk-conditions feature: wires repository →
- * service → router.
+ * service → routes.
  */
 export function createRiskConditionModule(prisma: PrismaService): DocumentedRouter {
   const repository = new RiskConditionRepository(prisma);
   const service = new RiskConditionService(repository);
-  return createRiskConditionRouter(service);
+  const doc = createDocumentedRouter();
+  registerRiskConditionRoutes(doc, service);
+  return doc;
 }
